+++ domain/model/schema_test.go
@@ -1,78 +1,10 @@
 package model
 
 import (
-	"strconv"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 )
-
-////// Table //////////////////////////
-
-func TestGenerateRecordsForTables(t *testing.T) {
-	type args struct {
-		vfc    map[ColumnFullName][]Value
-		schema Schema
-		n      int
-	}
-	tests := []struct {
-		name string
-		args args
-		want map[TableName][]Record
-	}{
-		{
-			name: "generate records for tables from map of column full-names and values",
-			args: args{
-				vfc: map[ColumnFullName][]Value{
-					"table1.test1": []Value{"1", "2", "3"},
-					"table1.test2": []Value{"aaa", "bbb", "ccc"},
-					"table2.test1": []Value{"v1", "v2", "v3"},
-				},
-				schema: Schema{
-					Tables: []Table{
-						{Name: "table1", Columns: []Column{{FullName: "table1.test1"}, {FullName: "table1.test2"}}},
-						{Name: "table2", Columns: []Column{{FullName: "table2.test1"}}},
-					},
-				},
-				n: 3,
-			},
-			want: map[TableName][]Record{
-				"table1": []Record{{"1", "aaa"}, {"2", "bbb"}, {"3", "ccc"}},
-				"table2": []Record{{"v1"}, {"v2"}, {"v3"}},
-			},
-		},
-		{
-			name: "skip auto increment columns",
-			args: args{
-				vfc: map[ColumnFullName][]Value{
-					"table1.test1": []Value{"1", "2", "3"},
-					"table1.test2": []Value{"aaa", "bbb", "ccc"},
-					"table2.test1": []Value{"v1", "v2", "v3"},
-				},
-				schema: Schema{
-					Tables: []Table{
-						{Name: "table1", Columns: []Column{{FullName: "table1.test1", AutoIncrement: true}, {FullName: "table1.test2"}}},
-						{Name: "table2", Columns: []Column{{FullName: "table2.test1"}}},
-					},
-				},
-				n: 3,
-			},
-			want: map[TableName][]Record{
-				"table1": []Record{{"aaa"}, {"bbb"}, {"ccc"}},
-				"table2": []Record{{"v1"}, {"v2"}, {"v3"}},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := GenerateRecordsForTables(tt.args.vfc, tt.args.schema, tt.args.n)
-			diff := cmp.Diff(got, tt.want)
-			if diff != "" {
-				t.Errorf("GenerateRecordsForTables(); -got, +want\n%v", diff)
-			}
-		})
-	}
-}
 
 ////// Column //////////////////////////
 
@@ -210,113 +142,6 @@ func TestColumn_GenerateData(t *testing.T) {
 			if diff != "" {
 				t.Error("Column.GenerateData(); -:got, +:want", diff)
 			}
-		})
-	}
-}
-
-func TestGenerateValuesForColumns(t *testing.T) {
-	type args struct {
-		cg SchemaGraph
-		n  int
-	}
-	tests := []struct {
-		name     string
-		args     args
-		assertFn func(map[ColumnFullName][]Value)
-	}{
-		{
-			name: "return map of values for columns considering foreign keys",
-			args: args{
-				cg: SchemaGraph{
-					AdjacencyMatrix: AdjacencyMatrix{
-						{0, 0, 0, 1},
-						{0, 0, 0, 0},
-						{0, 0, 0, 0},
-						{0, 1, 1, 0},
-					},
-					ColumnNodes: []ColumnNode{
-						ColumnNode{
-							column: Column{
-								FullName: "test0",
-								Type: ColumnType{
-									Base:  Int,
-									Param: ColumnTypeParam(1), //TODO: if the given schema is wrong(e.g. the param for column test0 was Int(3)), should sqloth validate it and notice users the inconsistency?
-								},
-							},
-							isDone: false,
-							index:  0,
-						},
-						ColumnNode{
-							column: Column{
-								FullName: "test1",
-								Type: ColumnType{
-									Base:  Int,
-									Param: ColumnTypeParam(1),
-								},
-							},
-							isDone: false,
-							index:  1,
-						},
-						ColumnNode{
-							column: Column{
-								FullName: "test2",
-								Type: ColumnType{
-									Base:  Int,
-									Param: ColumnTypeParam(1),
-								},
-							},
-							isDone: false,
-							index:  2,
-						},
-						ColumnNode{
-							column: Column{
-								FullName: "test3",
-								Type: ColumnType{
-									Base:  Int,
-									Param: ColumnTypeParam(2),
-								},
-							},
-							isDone: false,
-							index:  3,
-						},
-					},
-				},
-				n: 3,
-			},
-			assertFn: func(m map[ColumnFullName][]Value) {
-				for key, vs := range m {
-					switch key {
-					case "test0":
-						for idx, v := range vs {
-							if v != m["test3"][idx] {
-								t.Errorf("values of test1 is not valid; idx: %v, value: %v", idx, v)
-							}
-						}
-					case "test1", "test2":
-						for idx, v := range vs {
-							n, err := strconv.Atoi(string(v))
-							if err != nil {
-								t.Errorf("cannot convert value to int; value: %v", v)
-							}
-							if !(intRangeMap[Int][0] <= n || n <= intRangeMap[Int][1]) {
-								t.Errorf("values of %v is out of range; idx: %v, value: %v", key, idx, v)
-							}
-						}
-					case "test3":
-						for idx, v := range vs {
-							if v != m["test1"][idx]+m["test2"][idx] {
-								t.Errorf("values of test3 is not valid; idx: %v, value: %v", idx, v)
-							}
-						}
-					}
-				}
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := GenerateValuesForColumns(tt.args.cg, tt.args.n)
-			tt.assertFn(got)
 		})
 	}
 }
